refactor: re-organize Mapping and Operation modules, reformat
