+++ src/Data/Habulara.hs
@@ -2,4 +2,4 @@ module Data.Habulara
   ( module Core
   ) where
 
-import Data.Habulara.Core   as Core
+import Data.Habulara.Core as Core
